+++ test/test_client.py
@@ -2370,8 +2370,8 @@ class TestExhaustCursor(IntegrationTest):
 
     def setUp(self):
         super().setUp()
-        if client_context.is_mongos:
-            raise SkipTest("mongos doesn't support exhaust, SERVER-2627")
+        if not client_context.supports_exhaust_cursors():
+            raise SkipTest("mongos serves exhaust cursors only from 7.1, SERVER-57297")
 
     def test_exhaust_query_server_error(self):
         # When doing an exhaust query, the socket stays checked out on success